requirements.txt
fastapi
uvicorn
streamlit
pydantic
python-dotenv
openai
numpy
pandas
scikit-learn
requests
pytest